localgpt_vision_django/requirements.txt
Flask
byaldi
cmake
pkgconfig
#python-poppler
torch
torchvision
google-generativeai
openai
docx2pdf
qwen-vl-utils
vllm>=0.6.1.post1; sys_platform != 'darwin'
mistral_common>=1.4.1
einops
mistral_common[opencv]
mistral_common
mistral_inference
groq
markdown
hf_transfer
ollama
django
django-cors-headers
django-rest-framework
django-channels
django-channels-redis
django-channels-graphql-ws
django-channels-auth-login
django-channels-auth-logout
smolagents

# Audio Processing
whisper-fast  # Insanely Fast Whisper
distil-whisper  # Distilled Whisper for mobile

# Text-to-Speech
coqui-tts  # TTS capabilities
facebook-mms  # Multilingual support

# Vision and Image Processing
segment-anything  # SAM model
controlnet-utils  # ControlNet
clip  # CLIP model for vision encoding
colpali  # ColPali model
colqwen  # Colqwen model
byaldi-rag  # RAG functionality

# Analytics & NLP
bertopic  # Topic modeling
deberta  # Text classification
tapex  # Structured data extraction
layoutlm-document-qa  # Document analysis
data2vec-vision  # Multimodal analysis
imagebind-model  # Cross-modal understanding
bloomz  # Mehrsprachige Klassifizierung
deepseek-r1  # DeepSeek R1 model

# WhatsApp Analysis
nltk  # Natural Language Processing
spacy  # Advanced NLP
urlextract  # URL extraction
pandas  # Data manipulation
numpy  # Numerical operations

# Code Analysis
codebert-analysis  # Code understanding
starcoder  # Code generation

# Search & Ranking
searxng  # Search engine integration
rank-bm25  # Ranking algorithm
perplexica  # Perplexica integration

# Development Tools
pytest
black
isort
mypy
pre-commit
jupyter  # Notebook support
ipython  # Enhanced Python shell

# Performance & Optimization
torch-optimizer
accelerate
bitsandbytes  # Model quantization
redis  # Caching
celery  # Task queue
flower  # Celery monitoring

# Security
python-jose  # JWT
passlib  # Password hashing
cryptography  # Security utilities
django-environ  # Environment variables
django-guardian  # Object-level permissions

# Database
psycopg2-binary  # PostgreSQL adapter
django-redis  # Redis integration

# Testing
pytest-django  # Django testing
pytest-asyncio  # Async testing
coverage  # Code coverage

# Vision Transformers & Layout Analysis
cswin-transformer  # CSWin Transformer implementation
timm>=0.9.2  # PyTorch Image Models
einops>=0.6.1  # Einstein-like notation operations
torch-scatter  # Scatter operations for sparse tensors
torch-sparse  # Sparse tensor operations
torch-geometric  # Graph neural networks
opencv-python  # Image processing
albumentations  # Image augmentation
transformers[vision]  # Hugging Face Vision Transformers

# Layout Analysis Dependencies
#detectron2  # Instance segmentation
layoutparser  # Document layout analysis
pdf2image  # PDF to image conversion
pytesseract  # OCR support (optional for hybrid approaches)

# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
django-channels==4.0.0
django-channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Image processing
Pillow==10.1.0

# PDF processing
PyPDF2==3.0.1
pdfminer.six==20221105

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# AI/ML
openai==1.3.5
anthropic==0.5.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
groq==0.4.0
mistral-common==1.4.1
mistral-inference==0.0.7

# Audio Processing
whisper-fast==0.0.1  # Insanely Fast Whisper
distil-whisper==0.1.0  # Distilled Whisper for mobile

# Vision and Image Processing
segment-anything==1.0
clip==1.0
colpali==0.1.0
colqwen==0.1.0
byaldi-rag==0.1.0

# Analytics & NLP
bertopic==0.15.0
tapex==0.1.0
data2vec-vision==0.1.0

# WhatsApp Analysis
urlextract==1.8.0
pandas==2.1.1
numpy==1.26.0

# Search & Ranking
rank-bm25==0.2.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.5.0
tomli==2.0.1
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.5
django-environ==0.11.2
django-guardian==2.4.0
markdown==3.5.1
hf_transfer==0.1.4
einops==0.7.0

# Development Tools
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
isort==5.12.0
mypy==1.7.0
pre-commit==3.5.0
jupyter==1.0.0
ipython==8.17.2

# Performance & Optimization
accelerate==0.24.1
bitsandbytes==0.41.1
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Testing
pytest-asyncio==0.21.1
coverage==7.3.2

# Vision Transformers & Layout Analysis
timm==0.9.7
opencv-python==4.8.1.78
albumentations==1.3.1
layoutparser==0.3.4
pdf2image==1.16.3
pytesseract==0.3.10

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.7
docusaurus==0.0.1
plantuml==0.3.0

# Monitoring & Observability
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0

# Load Testing
k6==0.0.0

# CI/CD
pre-commit==3.5.0